requirements.txt
azure-storage-blob
azure-identity
azure-functions
pillow
IPTCInfo3
pymediainfo